+++ jsx/Filter.js
@@ -88,6 +88,9 @@ function Filter(props) {
           case 'date':
             element = <DateElement/>;
             break;
+          case 'time':
+            element = <TimeElement/>;
+            break;
           case 'checkbox':
             element = <CheckboxElement/>;
             break;