+++ webpack.config.js
@@ -290,6 +290,7 @@ const config = [
   lorisModule('server_processes_manager', ['server_processes_managerIndex']),
   lorisModule('instruments', ['CandidateInstrumentList']),
   lorisModule('candidate_profile', ['CandidateInfo']),
+  lorisModule('schedule_module', ['scheduleIndex']),
   lorisModule('api_docs', ['swagger-ui_custom']),
 ];